refactor: change PickupType dir location
