+++ src/main/java/com/si516/saludconecta/enums/PickupType.java
@@ -1,4 +1,4 @@
-package com.si516.saludconecta.document;
+package com.si516.saludconecta.enums;
 
 public enum PickupType {
     NOW,
